refacto StudyPosts buttons
