+++ src/components/study/buttons/StudyPostsButtons.module.css
@@ -1,0 +1,3 @@
+.buttons {
+  gap: 1rem;
+}
