+++ src/components/study/infography/StudyPostInfography.module.css
@@ -1,10 +1,3 @@
-.buttons {
-  position: absolute !important;
-  top: 0;
-  right: 0;
-  gap: 1rem;
-}
-
 .infography {
   margin-bottom: 2rem;
 }
